Move fake event clocks to testing subdirectory
